pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ddstartup"
version = "1.0.0"
description = "DD Startup: Fusion reactor tritium breeding and startup time analysis"
authors = [
    { name = "Alessandro Morandi" },
    { name = "Samuele Meschini" },
    { name = "Gabriele Iob" },
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "h5py>=3.0.0",
    "numba>=0.54.0",
    "joblib>=1.0.0",
    "tqdm>=4.60.0",
    "pyyaml>=5.4.0",
    "pint>=0.20.0",
    "plotly>=5.0.0",
    "seaborn>=0.11.0",
    "scikit-learn>=0.24.0",
    "SALib>=1.4.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Nuclear Fusion",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "black>=21.0",
    "flake8>=3.9.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
]

[project.scripts]
ddstartup = "src.main:main"
ddstartup-postprocess = "src.postprocessing.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.setuptools.package-data]
"src.registry" = ["*.yaml"]

# ---------- pytest ----------
[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "physics: marks tests related to physics calculations",
    "postprocessing: marks tests related to postprocessing functions",
]

# ---------- coverage ----------
[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/__init__.py",
    "setup.py",
    "*/conftest.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 0
precision = 2
